requirements.txt
baidu_aip==4.16.6
cnocr==2.1.2.1
colorama==0.4.5
Pillow==9.2.0
xlrd==1.2.0
